requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ctext==0.265
DateTime==5.1
dill==0.3.6
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.3.3
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
googletrans
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.14
nltk==3.8.1
openai==0.27.8
openlibrary==0.0.1
packaging==23.1
pathos==0.3.0
pluggy==1.2.0
pox==0.3.2
ppft==1.7.6.6
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyExecJS==1.5.1
pytest==7.4.0
pytz==2023.3
PyYAML==6.0
Random-Word==1.0.11
RandomWords==0.4.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
translate-api==4.9.5
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.6
win32-setctime==1.1.0
yarl==1.9.2
zope.interface==6.0